tests/wpt/web-platform-tests/referrer-policy/generic/iframe-upgrade-request.sub.html

<!DOCTYPE html>
<link rel="author" title="Dominic Farolino" href="dom@chromium.org">
<script src="https://{{domains[www]}}:{{ports[https][0]}}/resources/testharness.js"></script>
<script src="https://{{domains[www]}}:{{ports[https][0]}}/resources/testharnessreport.js"></script>
<body>
<iframe id="iframe"></iframe>
<script>
async_test(t => {
  const iframe = document.querySelector('iframe');
  const insecure_origin = new URL(location.href).origin;
  iframe.src = insecure_origin + '/referrer-policy4generic/resources/referrer.py';

  iframe.onload = t.step_func_done(() => {
    assert_true(iframe.contentDocument,
      "The iframe's contentDocument should be accessible, because the iframe");
    assert_equals(iframe.contentDocument.body.textContent, location.href + '/',
      "The referrer header sent for the iframe request should be redacted");
  });
}, "If an insecure iframe request is upgraded to https to be cross-origin, " +
   "referrer policies that consider same-origin-ness should be applied correctly");
</script>
</body>
